% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{models_general_create_datatable}
\alias{models_general_create_datatable}
\title{Create General Datatable}
\usage{
models_general_create_datatable(
  model_obj_tbl,
  datatable_title_chr,
  round = 3,
  ...
)
}
\arguments{
\item{model_obj_tbl}{Your model object tibble}

\item{datatable_title_chr}{What to title the datatable}

\item{round}{how many decimal places to found values in the datatable}

\item{...}{other args taken in by the `models_general_modify_tibble_colnames` function}
}
\value{
A datatable object
}
\description{
This function creates a datatable for other aspects of your modeling process, such as capturing best metrics
}
