% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{models_general_modify_tibble_colnames}
\alias{models_general_modify_tibble_colnames}
\title{Modify Column Names of a Tibble}
\usage{
models_general_modify_tibble_colnames(model_obj_tbl, ...)
}
\arguments{
\item{model_obj_tbl}{Your model object tibble}

\item{...}{other args}
}
\description{
This function is reused to rename columns in your output datatables to create more presentatble tables
This function is used internally and does not need to be called explicitly by the user
}
