% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{models_logistic_create_confusion_matrix}
\alias{models_logistic_create_confusion_matrix}
\title{Create Confusion Matrix}
\usage{
models_logistic_create_confusion_matrix(pred_tbl)
}
\arguments{
\item{pred_tbl}{A tibble with predicted values of your model}
}
\value{
A list
}
\description{
This function creates a confusion matrix for your model
}
