% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{models_logistic_create_summary_datatable}
\alias{models_logistic_create_summary_datatable}
\title{Create Summary Datatable}
\usage{
models_logistic_create_summary_datatable(
  model_fit,
  round = 3,
  type = "Summary",
  title_chr = NULL,
  ...
)
}
\arguments{
\item{model_fit}{Your model fit object}

\item{round}{how many decimal places to found values in the datatable}

\item{type}{default is "Summary"}

\item{title_chr}{default is NULL}

\item{...}{other args taken in by the `DT::datatable` function}
}
\value{
A datatable object
}
\description{
This function creates a summary datatable for your model
}
