% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tuning_eval}
\alias{tuning_eval}
\title{Evaluate Tuning Results}
\usage{
tuning_eval(tuning_obj)
}
\arguments{
\item{tuning_obj}{A tibble with a `TUNE_RESULTS` column containing `tune::tune_results` objects}
}
\value{
A tibble with added columns:
\describe{
  \item{TUNING_METRICS}{All evaluation metrics}
  \item{BEST_OF_METRICS}{Best results per metric}
  \item{BEST_OF_METRICS_TABLE}{Datatable summarizing best metrics}
  \item{BEST_OF_METRICS_PLOT}{Faceted ggplot for best metrics}
  \item{BEST_OF_METRICS_PLOTLY}{Interactive Plotly version of the plot}
}
}
\description{
This function processes a tuning object to extract the best model parameters and generate evaluation plots and tables.
}
